fix(cypress): update commands TS declaration to avoid errors
