+++ packages/e2e-cypress/src/helpers/commands.ts
@@ -1,5 +1,32 @@
-// Had to setup a separate file for the types, otherwise there were typescript/eslint errors in the editor.
-import './command-types';
+declare global {
+  namespace Cypress {
+    interface Chainable {
+      /**
+       * Custom command to select DOM element by data-cy attribute.
+       * @example cy.dataCy('greeting')
+       */
+      dataCy<E extends Node = HTMLElement>(value: string): Chainable<JQuery<E>>;
+
+      /**
+       * Custom command to test being on a given route.
+       * @example cy.testRoute('home')
+       */
+      testRoute(value: string): void;
+
+      /**
+       * Persist current local storage data.
+       * @example cy.saveLocalStorage()
+       */
+      saveLocalStorage(): void;
+
+      /**
+       * Restore saved data to local storage.
+       * @example cy.restoreLocalStorage()
+       */
+      restoreLocalStorage(): void;
+    }
+  }
+}
 
 export const registerCommands = () => {
   Cypress.Commands.add(
